+++ integration/redirects-test.ts
@@ -98,6 +98,34 @@ test.describe("redirects", () => {
             return <h1>Hello B!</h1>
           }
         `,
+
+        "app/routes/replace.a.tsx": js`
+          import { Link } from "@remix-run/react";
+          export default function () {
+            return <><h1 id="a">A</h1><Link to="/replace/b">Go to B</Link></>;
+          }
+        `,
+
+        "app/routes/replace.b.tsx": js`
+          import { Link } from "@remix-run/react";
+          export default function () {
+            return <><h1 id="b">B</h1><Link to="/replace/c">Go to C</Link></>
+          }
+        `,
+
+        "app/routes/replace.c.tsx": js`
+          import { replace } from "@remix-run/node";
+          export const loader = () => replace("/replace/d");
+          export default function () {
+            return <h1 id="c">C</h1>
+          }
+        `,
+
+        "app/routes/replace.d.tsx": js`
+          export default function () {
+            return <h1 id="d">D</h1>
+          }
+        `,
       },
     });
 
@@ -142,6 +170,18 @@ test.describe("redirects", () => {
     await page.waitForSelector('h1:has-text("Hello B!")');
     // Hard reload resets client side react state
     expect(await app.getHtml("button")).toMatch("Count:0");
+  });
+
+  test("supports replace redirects within the app", async ({ page }) => {
+    let app = new PlaywrightFixture(appFixture, page);
+    await app.goto("/replace/a", true);
+    await page.waitForSelector("#a"); // [/a]
+    await app.clickLink("/replace/b");
+    await page.waitForSelector("#b"); // [/a, /b]
+    await app.clickLink("/replace/c");
+    await page.waitForSelector("#d"); // [/a, /d]
+    await page.goBack();
+    await page.waitForSelector("#a"); // [/a]
   });
 });
 
@@ -243,6 +283,34 @@ test.describe("single fetch", () => {
               return <h1>Hello B!</h1>
             }
           `,
+
+          "app/routes/replace.a.tsx": js`
+            import { Link } from "@remix-run/react";
+            export default function () {
+              return <><h1 id="a">A</h1><Link to="/replace/b">Go to B</Link></>;
+            }
+          `,
+
+          "app/routes/replace.b.tsx": js`
+            import { Link } from "@remix-run/react";
+            export default function () {
+              return <><h1 id="b">B</h1><Link to="/replace/c">Go to C</Link></>
+            }
+          `,
+
+          "app/routes/replace.c.tsx": js`
+            import { replace } from "@remix-run/node";
+            export const loader = () => replace("/replace/d");
+            export default function () {
+              return <h1 id="c">C</h1>
+            }
+          `,
+
+          "app/routes/replace.d.tsx": js`
+            export default function () {
+              return <h1 id="d">D</h1>
+            }
+          `,
         },
       });
 
@@ -289,6 +357,18 @@ test.describe("single fetch", () => {
       await page.waitForSelector('h1:has-text("Hello B!")');
       // Hard reload resets client side react state
       expect(await app.getHtml("button")).toMatch("Count:0");
+    });
+
+    test("supports replace redirects within the app", async ({ page }) => {
+      let app = new PlaywrightFixture(appFixture, page);
+      await app.goto("/replace/a", true);
+      await page.waitForSelector("#a"); // [/a]
+      await app.clickLink("/replace/b");
+      await page.waitForSelector("#b"); // [/a, /b]
+      await app.clickLink("/replace/c");
+      await page.waitForSelector("#d"); // [/a, /d]
+      await page.goBack();
+      await page.waitForSelector("#a"); // [/a]
     });
   });
 });