+++ packages/remix-server-runtime/routes.ts
@@ -7,7 +7,6 @@ import type {
 import { callRouteAction, callRouteLoader } from "./data";
 import type { FutureConfig } from "./entry";
 import type { ServerRouteModule } from "./routeModules";
-import type { DataStrategyCtx } from "./single-fetch";
 
 export interface RouteManifest<Route> {
   [routeId: string]: Route;
@@ -101,8 +100,6 @@ export function createStaticHandlerDataRoutes(
               loader: route.module.loader!,
               routeId: route.id,
               singleFetch: future.unstable_singleFetch === true,
-              response: (dataStrategyCtx as DataStrategyCtx | undefined)
-                ?.response,
             })
         : undefined,
       action: route.module.action
@@ -114,8 +111,6 @@ export function createStaticHandlerDataRoutes(
               action: route.module.action!,
               routeId: route.id,
               singleFetch: future.unstable_singleFetch === true,
-              response: (dataStrategyCtx as DataStrategyCtx | undefined)
-                ?.response,
             })
         : undefined,
       handle: route.module.handle,